Add files via upload
Attached is the update data as of 05/07. As per Prof. Lai's comments during the presentation, median values are also included. We should not fit a model using both "Mean_comp_count" and "Mean_comp_count_alt", as these are separate variables. The same is true for "Mean_comp_count" and "Med_comp_count_alt".  We are trying to predict "Class".